requirements.txt
pillow
pytesseract
pyinstaller
